Find frame files using name filters
Without this commit, nvdu_viz always looks for '*.png' frame files. This
means that it always fails for datasets where the frame files are named
'*.jpg', such as the Falling Things dataset.

This commit respects the name filters, so that the Falling Things
dataset can now be visualized like this:

    nvdu_viz --name_filters '*.jpg'